lift state of fetchData
